requirements.txt
duckdb
pandas
scikit-learn
matplotlib
python-dotenv
fastapi
uvicorn
requests
pytest
httpx

